Move `cloneIfNecessary()` to `value-clone.ts`.
